+++ src/screen/doctor/index.tsx
@@ -1,10 +1,10 @@
 import { useState, useEffect } from "react";
-import TabBar from "../../components/TabBar/buttonTabBar";
-import DataDoctor from "../../components/Patients/dataDoctor/data-doctor";
+import TabBar from "../../components/TabBar";
+import DataDoctor from "../../components/Patients/dataDoctor";
 import {
   ContainerSafePrincipalPaciente,
   Scroll,
-} from "./profissionalResponsavel-style";
+} from "./style";
 import Header from "../../components/Header"
 import React from "react";
 import storePatient from "@/src/store/storePatient";
@@ -16,16 +16,20 @@ const Doctor = () => {
     <ContainerSafePrincipalPaciente>
       <Scroll>
         <Header
-          title={patient.doctorInfo.name}
+          // title={patient.doctorInfo.name}
+          title={'Juliana nogueira'}
           backButton
           vaccineButton
-          photo={patient.doctorInfo.photo}
+          // photo={patient.doctorInfo.photo}
           type='doctor'
         />
         <DataDoctor
-          crm={patient.doctorInfo.CRM}
-          sobre={patient.doctorInfo.about}
-          especialidade={patient.doctorInfo.specialty}
+          // crm={patient.doctorInfo.CRM}
+          // sobre={patient.doctorInfo.about}
+          // especialidade={patient.doctorInfo.specialty}
+          crm={'123'}
+          sobre="sobre"
+          especialidade="dentista"
         />
       </Scroll>
       <TabBar />