+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Changelog
 - [PADA-189](https://givisiez.atlassian.net/browse/PADA-189): Create Medication Controll Header Layout
 - [PADA-210](https://givisiez.atlassian.net/browse/PADA-210): Create date filter
 - [PADA-244](https://givisiez.atlassian.net/browse/PADA-244): Add password toggle function
+- [PADA-254](https://givisiez.atlassian.net/browse/PADA-254): Create edit profile and changing password. Add passage routes
+- [PADA-283](https://givisiez.atlassian.net/browse/PADA-283): Add font imports at app.json and app.tsx
+- [PADA-285](https://givisiez.atlassian.net/browse/PADA-285): Add style themes archives
 - [PADA-283](https://givisiez.atlassian.net/browse/PADA-283): Add font imports at app.json and app.tsx
 - [PADA-285](https://givisiez.atlassian.net/browse/PADA-285): Add style themes archives
 ## Fix